chore: move author profile provider to author profile
