+++ packages/author-profile/author-profile-query.js
@@ -1,0 +1,46 @@
+import gql from "graphql-tag";
+
+export default gql`
+  query ArticleQuery(
+    $slug: Slug!
+    $first: Int
+    $skip: Int
+    $imageRatio: Ratio!
+  ) {
+    author(slug: $slug) {
+      name
+      jobTitle
+      biography
+      image
+      twitter
+      articles {
+        count
+        list(first: $first, skip: $skip) {
+          content(maxCharCount: 145, markupType: "paragraph")
+          id
+          label
+          url
+          leadAsset {
+            ... on Image {
+              title
+              crop(ratio: $imageRatio) {
+                url
+              }
+            }
+            ... on Video {
+              posterImage {
+                title
+                crop(ratio: $imageRatio) {
+                  url
+                }
+              }
+            }
+          }
+          publicationName
+          publishedTime
+          title
+        }
+      }
+    }
+  }
+`;
