+++ packages/author-profile/author-profile.stories.js
@@ -2,7 +2,13 @@ import React from "react";
 import { StyleSheet, View } from "react-native";
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { storiesOf } from "@storybook/react-native";
-import AuthorProfile from "./author-profile";
+import {
+  ApolloClient,
+  ApolloProvider,
+  createNetworkInterface,
+  IntrospectionFragmentMatcher
+} from "react-apollo";
+import Provider, { Component as AuthorProfile } from "./author-profile";
 import example from "./example.json";
 
 const styles = StyleSheet.create({
@@ -23,8 +29,38 @@ const story = m =>
     </View>
   </View>;
 
-storiesOf("AuthorProfile", module)
-  .add("AuthorProfile", () => {
+const fragmentMatcher = new IntrospectionFragmentMatcher({
+  introspectionQueryResultData: {
+    __schema: {
+      types: [
+        {
+          kind: "UNION",
+          name: "Media",
+          possibleTypes: [
+            {
+              name: "Image"
+            },
+            {
+              name: "Video"
+            }
+          ]
+        }
+      ]
+    }
+  }
+});
+
+const networkInterface = createNetworkInterface({
+  uri: "http://localhost:4000/graphql/"
+});
+
+const client = new ApolloClient({
+  networkInterface,
+  fragmentMatcher
+});
+
+storiesOf("Author Profile", module)
+  .add("Component", () => {
     const props = {
       data: Object.assign({}, example, {
         count: example.articles.count,
@@ -41,17 +77,24 @@ storiesOf("AuthorProfile", module)
 
     return story(<AuthorProfile {...props} />);
   })
-  .add("AuthorProfile Loading", () => {
+  .add("Loading", () => {
     const props = {
       isLoading: true
     };
 
     return story(<AuthorProfile {...props} />);
   })
-  .add("AuthorProfile Empty State", () => {
+  .add("Empty", () => {
     const props = {
       isLoading: false
     };
 
     return story(<AuthorProfile {...props} />);
-  });
+  })
+  .add("Provider", () =>
+    story(
+      <ApolloProvider client={client}>
+        <Provider slug="fiona-hamilton" imageRatio="3:2" />
+      </ApolloProvider>
+    )
+  );
