+++ packages/provider/package.json
@@ -52,7 +52,6 @@
     "webpack": "3.3.0"
   },
   "dependencies": {
-    "graphql-tag": "2.4.2",
     "lodash.get": "4.4.2",
     "lodash.pick": "4.4.0",
     "prop-types": "15.5.10",
